moved nav-items file into constants folder
