+++ src/frontend/components/navigation/mobile-nav.tsx
@@ -12,7 +12,7 @@ import { ViewVerticalIcon } from '@radix-ui/react-icons';
 import { Button } from '../ui/button';
 import { ScrollArea } from '../ui/scroll-area';
 
-import navItems from '@/frontend/components/navigation/nav-items';
+import navItems from '@/frontend/constants/nav-items';
 
 const MobileNav = () => {
 	const [open, setOpen] = React.useState(false);
