+++ src/frontend/constants/nav-items.ts
